requirements.txt
# Data Science and ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-optimize>=0.9.0
xgboost>=2.0.0
tensorflow>=2.13.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
kaggle>=1.5.16

# Web Framework (for future web interface)
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0
tqdm>=4.65.0
PyYAML>=6.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
jupyter>=1.0.0

# GUI (tkinter comes with Python, but these are optional enhancements)
# pillow>=10.0.0  # For image processing in GUI
# ttkbootstrap>=1.10.1  # For modern GUI themes (optional)

# Database (sqlite3 is built-in to Python)
# sqlite3  # Built-in module, no installation needed
